feat: add Validation and Exception at Channel Service
