+++ src/main/java/com/sprint/mission/discodeit/mapper/entitymapper/ChannelMapper.java
@@ -3,9 +3,12 @@ package com.sprint.mission.discodeit.mapper.entitymapper;
 import com.sprint.mission.discodeit.dto.response.ChannelDto;
 import com.sprint.mission.discodeit.dto.response.UserDto;
 import com.sprint.mission.discodeit.entity.channel.Channel;
+import com.sprint.mission.discodeit.exception.ErrorCode;
 import com.sprint.mission.discodeit.exception.channel.ChannelNotFoundException;
 import com.sprint.mission.discodeit.repository.ReadStatusRepository;
+import java.time.Instant;
 import java.util.List;
+import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -17,7 +20,9 @@ public class ChannelMapper {
 
   public ChannelDto toDto(Channel channel) {
     if (channel == null) {
-      throw new ChannelNotFoundException();
+      throw new ChannelNotFoundException(Instant.now(), ErrorCode.CHANNEL_NOT_FOUND,
+          Map.of(ErrorCode.CHANNEL_NOT_FOUND.getCode(), ErrorCode.CHANNEL_NOT_FOUND.getMessage())
+      );
     }
 
     List<UserDto> users = readStatusRepository.findAllByChannel(channel).stream()
